CODE-209 Split up large menu class to be composed of sub-classes
