+++ sous-chef/sous_chef/grocery_list/generate_grocery_list.py
@@ -11,7 +11,10 @@ from sous_chef.formatter.format_str import convert_number_to_str
 from sous_chef.formatter.format_unit import UnitFormatter, unit_registry
 from sous_chef.formatter.ingredient.format_ingredient import Ingredient
 from sous_chef.formatter.ingredient.get_ingredient_field import IngredientField
-from sous_chef.menu.create_menu import MenuIngredient, MenuRecipe
+from sous_chef.menu.create_menu._for_grocery_list import (
+    MenuIngredient,
+    MenuRecipe,
+)
 from sous_chef.messaging.todoist_api import TodoistHelper
 from sous_chef.recipe_book.read_recipe_book import Recipe
 from structlog import get_logger
