+++ sous-chef/sous_chef/grocery_list/main.py
@@ -6,7 +6,7 @@ from sous_chef.formatter.format_unit import UnitFormatter
 from sous_chef.formatter.ingredient.format_ingredient import IngredientFormatter
 from sous_chef.formatter.ingredient.get_ingredient_field import IngredientField
 from sous_chef.grocery_list.generate_grocery_list import GroceryList
-from sous_chef.menu.create_menu import Menu
+from sous_chef.menu.create_menu.create_menu import Menu
 from sous_chef.messaging.gsheets_api import GsheetsHelper
 from sous_chef.messaging.todoist_api import TodoistHelper
 from sous_chef.pantry_list.read_pantry_list import PantryList
