+++ sous-chef/sous_chef/menu/create_menu/create_menu.py
@@ -1,0 +1,17 @@
+from dataclasses import dataclass
+from pathlib import Path
+
+from sous_chef.menu.create_menu._for_grocery_list import MenuForGroceryList
+from sous_chef.menu.create_menu._for_todoist import MenuForTodoist
+from sous_chef.menu.create_menu._from_fixed_template import (
+    MenuFromFixedTemplate,
+)
+from structlog import get_logger
+
+ABS_FILE_PATH = Path(__file__).absolute().parent
+FILE_LOGGER = get_logger(__name__)
+
+
+@dataclass
+class Menu(MenuForTodoist, MenuForGroceryList, MenuFromFixedTemplate):
+    pass
